Issue 62, Flash Calibration sector must be reserved in ld linker script.
This should be done to ensure that write of calibration data will not
clear sections of the program data (.text etc).

+++ STM32/DC/Core/Src/main.c
@@ -146,7 +146,7 @@ void printResult(int16_t *pBuffer, int noOfChannels, int noOfSamples)
     if (isFirstWrite)
         clearLineAndBuffer();
 
-    double temp;
+    float temp;
     if (si7051Temp(&hi2c1, &temp) != HAL_OK)
         temp = 10000;
